// src/app/03_SIR.Produccion.Reportes/pages/cuota/interfaces/DWSalidaDTO.interface.ts
export interface DWSalidaDTO {
    fecha:          Date;
    lote:           number;
    customer?:      string;
    condition:      string;
    marca:          number;
    material:       number;
    codigo:         string;
    producto:       string;
    piezas:         number;
    peso:           number;
}